% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycoverse-package.R
\docType{package}
\name{glycoverse-package}
\alias{glycoverse}
\alias{glycoverse-package}
\title{glycoverse: Easy Install and Load the `Glycoverse`}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The 'glycoverse' is a set of packages that together form a comprehensive pipeline for glycomics and glycoproteomics data analysis. This package is designed to make it easy to install and load multiple 'glycoverse' packages in a single step.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://glycoverse.github.io/glycoverse/}
}

}
\author{
\strong{Maintainer}: Bin Fu \email{23110220018@m.fudan.edu.cn} (\href{https://orcid.org/0000-0001-8567-2997}{ORCID})

}
\keyword{internal}
