% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{glycoverse_deps}
\alias{glycoverse_deps}
\title{List all glycoverse dependencies}
\usage{
glycoverse_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
glycoverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\description{
List all glycoverse dependencies
}
