% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{glycoverse_packages}
\alias{glycoverse_packages}
\title{List all packages in the glycoverse}
\usage{
glycoverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include glycoverse in the list?}
}
\description{
List all packages in the glycoverse
}
\examples{
glycoverse_packages()
}
