% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{glycoverse_sitrep}
\alias{glycoverse_sitrep}
\title{Get a situation report on the glycoverse}
\usage{
glycoverse_sitrep()
}
\description{
This function gives a quick overview of the versions of R and RStudio as
well as all glycoverse packages. It's primarily designed to help you get
a quick idea of what's going on when you're helping someone else debug
a problem.
}
