% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{glycoverse_update}
\alias{glycoverse_update}
\title{Update glycoverse packages}
\usage{
glycoverse_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
glycoverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\description{
This checks r-universe for all glycoverse packages (and optionally,
their dependencies) and reports which ones have newer versions available.
You can then choose to update automatically or get manual instructions.
}
\examples{
\dontrun{
glycoverse_update()
}
}
